pyproject.toml
[tool.poetry]
name = "wildberries-bot-team-1"
version = "0.1.0"
description = ""
authors = ["Сергей Шубин", "Сергей Парфенов", "Андрей Куликов", "Анна Горбунова", "Артемий Березин", "Кирилл Молчанов", "Яна Денисова"]
readme = "README.md"
packages = [{include = "wildberries_bot_team_1"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
python-telegram-bot = {extras = ["job-queue"], version = "^20.3"}
flake8-pyproject = "^1.2.3"
aiohttp = "^3.8.4"
pydantic = "^1.10.9"
django-phonenumber-field = "^7.1.0"
prettytable = "^3.8.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
autopep8 = "^2.0.2"
pre-commit = "^3.3.2"
add-trailing-comma = "^2.4.0"
mypy = "^1.3.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
exclude = ".venv"
paths = ["./tgbot", "./bot", "./botmanager"]
max-complexity = 10
max-line-length = 79
per-file-ignores = """
    __init__.py: F401
    settings.py: E501
"""

[tool.isort]
default_section = "THIRDPARTY"
sections = ["STDLIB", "FIRSTPARTY", "THIRDPARTY", "LOCALFOLDER"]
known_local_folder = ["tgbot", "bot", "botmanager"]
lines_after_imports = "2"
skip = ["tests", ".venv", "venv"]
line_length = "79"
multi_line_output = "3"
include_trailing_comma = "true"
use_parentheses = "true"
